Ensure all files are in job_applications and not job_posts
