+++ client/src/utils/job_posts_api.js
@@ -137,35 +137,6 @@ const getAllLocations = async () => {
   return response;
 };
 
-const getFilterOptions = async (queryParams = {}) => {
-  // make a query string from given parameters, no queryString if empty
-  const queryString = new URLSearchParams(queryParams).toString();
-
-  try {
-    const response = await fetch(
-      `${REACT_APP_API_BASE_URL}/job_postings/filter-options?${queryString}`,
-      {
-        method: "GET",
-        credentials: "include",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      },
-    );
-
-    if (!response.ok) {
-      const error = await response.json();
-      throw new Error(`Error retrieving filter options: ${error.message}`);
-    }
-
-    const result = await response.json();
-    return result.data;
-  } catch (error) {
-    console.error("Error fetching filter options:", error);
-    throw error;
-  }
-};
-
 module.exports = {
   createJobPost,
   deleteJobPost,
@@ -175,5 +146,4 @@ module.exports = {
   getOneJobPost,
   getAllLocations,
   getOneActiveJobPost,
-  getFilterOptions,
 };
