+++ server/src/controllers/job_applications/getFilterDropdownOptionsHandler.js
