+++ server/src/routes/job_posts.js
@@ -17,7 +17,6 @@ const deleteJobPostHandler = require("../controllers/job_posts/deleteJobPost");
 const isLoggedIn = require("../middlewares/auth/isLoggedIn");
 
 const getAllLocationsRequestHandler = require("../controllers/job_posts/getAllLocations");
-const getFilterDropdownOptionsHandler = require("../controllers/job_posts/getFilterDropdownOptionsHandler");
 
 const getActiveJobPostRequestHandler = require("../controllers/job_posts/getOneActiveJobPost");
 
@@ -183,24 +182,3 @@ router.put("/:job_posting_id", isLoggedIn, updateJobPostRequestHandler);
  * - isLoggedIn: Ensures that the user is authenticated before allowing the deletion operation.
  */
 router.delete("/:job_posting_id", isLoggedIn, deleteJobPostHandler);
-
-/**
- * Get Filter Options for Dropdowns
- *
- * Retrieves unique applicant names, emails, job titles, and job posting IDs.
- * Supports filtering by applicant name, email, job title, and job posting ID.
- *
- * @type GET
- *
- * @route /filter-options
- *
- * Query Parameters:
- * @param {string} [applicant_name]
- * @param {string} [email]
- * @param {string} [job_title]
- * @param {string} [job_posting_id]
- */
-
-router.get("/filter-options", isLoggedIn, getFilterDropdownOptionsHandler);
-
-module.exports = router;
